Fix catastrophic cancellation of derivative terms for vapoil model.
In terms of derivatives, we ended up writing a = a - a, removing the
derivatives of the oil equation w.r.t. the water saturation when there
was no oil. This fixes CPR for SPE3 for new equation ordering (and will
maybe improve convergence on other vapoil cases).

+++ ad-blackoil/utils/calculateHydrocarbonsFromStatusBO.m
@@ -1,6 +1,8 @@
 function [sG, rs, rv, rsSat, rvSat] = calculateHydrocarbonsFromStatusBO(model, status, sO, x, rs, rv, pressure)
     % define sG, rs and rv in terms of x
-    sG = status{2}.*sO + status{3}.*x;
+    % We cast "sO" to double to avoid cancellation of derivatives when
+    % calculating the final sO later.
+    sG = status{2}.*double(sO) + status{3}.*x;
     if model.disgas
         rsSat = model.fluid.rsSat(pressure);
         rs = (~status{1}).*rsSat + status{1}.*x;
